Restructure data and components
