+++ .gitignore
@@ -3,4 +3,3 @@ dist/
 dist-es/
 stats.json
 report.html
-/src/data/data.js
